Fixed crash due to int overflow.

+++ slinest.h
@@ -7,7 +7,7 @@
 struct slinest {
 	uint8_t *sl_text;	/* the text of the line */
 	uint32_t *sl_rend;	/* the rendition style */
-	uint16_t sl_length;	/* length of the line */
+	uint8_t sl_length;	/* length of the line */
 };
 
 #endif//!SLINEST_H
